Style Modify & Delete Post buttons, add confirm on logout fx and delete post fx

+++ client-react/src/components/SignOutComponent.js
@@ -11,7 +11,7 @@ const StyledLink = styled(Link)`
   transition: background-color 0.2s ease;
   &:hover {
     background-color: var(--primary);
-    color: #F1F1F1;
+    color: #f1f1f1;
   }
 `;
 
@@ -19,9 +19,16 @@ export default function SignOutComponent() {
   const navigate = useNavigate();
 
   const logout = () => {
-    localStorage.clear();
-    navigate("/", { replace: true });
-    window.location.reload();
+    let result = window.confirm("Voulez-vous vous déconnecter ?");
+
+    if (result === true) {
+      localStorage.clear();
+      navigate("/", { replace: true });
+      window.location.reload();
+    }
+    else {
+      return
+    }
   };
   return <StyledLink onClick={logout}>Se déconnecter</StyledLink>;
 }
